requirements.txt
# Space Station - Python Dependencies

# Astronomy
astropy>=6.0
skyfield>=1.46
sgp4>=2.22

# SDR
pyrtlsdr>=0.3.0

# Signal Processing
numpy>=1.24
scipy>=1.11
matplotlib>=3.7

# Web interface
fastapi>=0.104
uvicorn>=0.24
websockets>=12.0
jinja2>=3.1

# Hardware control (RPi)
# pigpio>=1.78  # Install via apt: sudo apt install pigpio python3-pigpio
# smbus2>=0.4   # I2C for AS5600 encoders
RPi.GPIO>=0.7; platform_machine == "aarch64" or platform_machine == "armv7l"

# Configuration
pyyaml>=6.0

# Data
pandas>=2.0

# Utilities
click>=8.1      # CLI interface
rich>=13.0      # Pretty terminal output
